--- frontend/src/Component/banner/index.css
@@ -1,30 +1,0 @@
-.banner-container{
-    width: 100%;
-}
-.banner-description{
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    text-align: center;
-}
-.banner{
-    width: 100%;
-    height: 150px;
-    border: 1px solid black;
-    margin: 20px 0 20px 0;
-}
-.bannerimg{
-    width: 450px;
-    height: 300px;
-    margin: 10px 0 10px 0;
-    box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.6);
-    border: 3px solid #289aff;
-    border-radius: 10px;
-}
-.banner-description_p h2{
-    color: #289aff;
-}
-.banner-description_p_box{
-    border: 1px solid black;
-    padding: 0 5px 0 5px;
-}
